+++ src/Exception.php
@@ -5,10 +5,12 @@
  * @package    League\Uri
  * @subpackage League\Uri
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
- * @copyright  2016 Ignace Nyamagana Butera
  * @license    https://github.com/thephpleague/uri-parser/blob/master/LICENSE (MIT License)
- * @version    1.0.0
+ * @version    1.0.5
  * @link       https://github.com/thephpleague/uri-parser/
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 declare(strict_types=1);